// src/common/enums/StatusResponse.enum.ts
export enum StatusResponse {
  SUCCESS = 1,
  FAIL = 2,
  FOBBIDEN = 3,
  NOTFOUND = 4,
  EXISTS_USERNAME = 5,
  EXISTS_EMAIL = 6,
  NOT_EXISTS_ROLE = 8,
  USERNAME_OR_PASSWORD_IS_NOT_CORRECT = 9,
  PASSWORD_INCORRECT = 17,
  NOT_EXIST_EMAIL = 18,
  WRONG_OTP = 19,
  WRONG_HASH_CODE = 20,
  DUPLICATE_DOMAIN = 21,
  NOT_EXISTS_DOMAIN = 22,
  NOT_EXISTS_LINK = 23,
  USERNAME_OR_PASSWORD_DOMAIN_IS_NOT_CORRECT = 24,
  DOMAIN_EXIST = 25,
  USER_BY_ID_NOT_FOUND = 26,
  NOT_FOUND = 27,
  TEAM_NAME_EXISTS = 28,
  NOT_EXISTS_TEAM = 29,
}